feat: Completely refactor the `FrameProcessorPlugins.ts` file into multiple files
